+++ src/backend/marsha/bbb/tests/api/classroomdocument/test_initiate_upload.py
@@ -58,7 +58,7 @@ class ClassroomDocumentInitiateUploadAPITest(TestCase):
             mock_dt.utcnow = mock.Mock(return_value=now)
             response = self.client.post(
                 f"/api/classroomdocuments/{classroom_document.id}/initiate-upload/",
-                {"filename": "foo.pdf", "mimetype": "application/pdf"},
+                {"filename": "foo.pdf", "mimetype": "application/pdf", "size": 10},
                 HTTP_AUTHORIZATION=f"Bearer {jwt_token}",
                 content_type="application/json",
             )
@@ -116,7 +116,7 @@ class ClassroomDocumentInitiateUploadAPITest(TestCase):
             mock_dt.utcnow = mock.Mock(return_value=now)
             response = self.client.post(
                 f"/api/classroomdocuments/{classroom_document.id}/initiate-upload/",
-                {"filename": "foo", "mimetype": "application/pdf"},
+                {"filename": "foo", "mimetype": "application/pdf", "size": 10},
                 HTTP_AUTHORIZATION=f"Bearer {jwt_token}",
                 content_type="application/json",
             )
@@ -174,7 +174,7 @@ class ClassroomDocumentInitiateUploadAPITest(TestCase):
             mock_dt.utcnow = mock.Mock(return_value=now)
             response = self.client.post(
                 f"/api/classroomdocuments/{classroom_document.id}/initiate-upload/",
-                {"filename": "foo", "mimetype": ""},
+                {"filename": "foo", "mimetype": "", "size": 10},
                 HTTP_AUTHORIZATION=f"Bearer {jwt_token}",
                 content_type="application/json",
             )
@@ -204,7 +204,7 @@ class ClassroomDocumentInitiateUploadAPITest(TestCase):
             mock_dt.utcnow = mock.Mock(return_value=now)
             response = self.client.post(
                 f"/api/classroomdocuments/{classroom_document.id}/initiate-upload/",
-                {"filename": "foo", "mimetype": "application/wrong-type"},
+                {"filename": "foo", "mimetype": "application/wrong-type", "size": 10},
                 HTTP_AUTHORIZATION=f"Bearer {jwt_token}",
                 content_type="application/json",
             )
@@ -236,7 +236,7 @@ class ClassroomDocumentInitiateUploadAPITest(TestCase):
             mock_dt.utcnow = mock.Mock(return_value=now)
             response = self.client.post(
                 f"/api/classroomdocuments/{classroom_document.id}/initiate-upload/",
-                {"filename": "foo.pdf", "mimetype": "application/pdf"},
+                {"filename": "foo.pdf", "mimetype": "application/pdf", "size": 10},
                 HTTP_AUTHORIZATION=f"Bearer {jwt_token}",
                 content_type="application/json",
             )
@@ -265,7 +265,7 @@ class ClassroomDocumentInitiateUploadAPITest(TestCase):
             mock_dt.utcnow = mock.Mock(return_value=now)
             response = self.client.post(
                 f"/api/classroomdocuments/{classroom_document.id}/initiate-upload/",
-                {"filename": "foo.pdf", "mimetype": "application/pdf"},
+                {"filename": "foo.pdf", "mimetype": "application/pdf", "size": 10},
                 HTTP_AUTHORIZATION=f"Bearer {jwt_token}",
                 content_type="application/json",
             )
@@ -326,7 +326,7 @@ class ClassroomDocumentInitiateUploadAPITest(TestCase):
             mock_dt.utcnow = mock.Mock(return_value=now)
             response = self.client.post(
                 f"/api/classroomdocuments/{classroom_document.id}/initiate-upload/",
-                {"filename": "foo.pdf", "mimetype": "application/pdf"},
+                {"filename": "foo.pdf", "mimetype": "application/pdf", "size": 10},
                 HTTP_AUTHORIZATION=f"Bearer {jwt_token}",
                 content_type="application/json",
             )